DZ2-13/hw_rest_api/pyproject.toml
[tool.poetry]
name = "hw-rest-api"
version = "0.1.0"
description = ""
authors = ["NikYurchik <14urka@gmail.com>"]
readme = "README.md"
packages = [{include = "hw_rest_api"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi-limiter = "^0.1.5"
pydantic-settings = "^2.0.3"
pydantic = {extras = ["settings"], version = "^2.3.0"}
python-dotenv = "^1.0.0"
cloudinary = "^1.34.0"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.20"
fastapi-mail = "^1.4.1"
fastapi-jwt-auth = "^0.5.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
libgravatar = "^1.0.4"
psycopg2 = "^2.9.7"
python-multipart = "^0.0.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
